requirements.txt
reflex==0.5.9
reflex-type-animation==0.0.1
feedparser==6.0.11
python-dotenv==1.0.1
pytz